+++ app/subcommands/config/add.sh
@@ -1,5 +1,6 @@
 #!/bin/sh
 # Description: Add a value to a list config key
+# Usage: <KEY> <VALUE>
 # vim: ft=sh ts=4 sw=4 sts=4 noet
 set -euf