+++ app/subcommands/config/get.sh
@@ -1,5 +1,6 @@
 #!/bin/sh
 # Description: Get the value assigned to the given config key
+# Usage: <KEY>
 # vim: ft=sh ts=4 sw=4 sts=4 noet
 set -euf